Expand scope for each mark multiple times in a row
NOTE: I haven't actually found any official documentation for the `edit`
object and associated commands, so I presume the behaviour is subject to
change at any time and this whole thing could break terribly.

However, in order for sublime to NOT batch up multiple of the mark
commands together I stop the original edit and begin a new one in its
place but with an additional UUID arg so that repeated invocations don't
match wholesale. If I don't do this I get really inconsistent behaviour
with the repeat count, so hopefully if this behaviour is changed, we get
a more solid/consistent way to detect repeated commands.

There may also already be such an API but if there is I can't find it 😞