Nicer "preview" of marked scope
Open two "temporary" views, one for the start and one for the end that
update as we set/increase the scope. Dismiss them as soon as the user
does something i.e. modifies the text, or performs a `text_command`.

This won't cover _absolutely every interaction_ that the user might do,
but it covers the majority and feels pretty nice so far.

Ideally we'd also hook into the "ESC" key, but from my initial research
that looks like some epic shenanigans that don't feel worth the effort
right now. (This was already crazier than I'd hoped it would be).